pose/requirements.txt
django
djangorestframework
django-cors-headers
opencv-python
mediapipe
numpy
matplotlib
fastdtw
gunicorn
torch>=0.4.1
torchvision>=0.2.1
pycocotools
